added in parsing arguments
Added in code to parse arguments in the command line as well as a help
section for parsing arguments. I also changed so that input and output
directories are specified. Also added in more spacers so that it is
easier to find descriptions of what each section of code does and
removed all obsolete code.