A bunch of bug fixes
Changed the name and letter of the arguments so that a clash can be
avoided. Also fixed some bugs line 120. Removed some obsolete commands.
Fixed a bug which was causing a lot of problems which was not changed
when i changed the name of the arguments to be passed. And finally
added in commands to run the trimming for all files in each folder.