added in parsing option for velvetg options.
There are too many options to tell velvetg to add in a specific parsing
argument for each option. Therefore you now just have one long string
of commands for velvetg. This simplifies the coding but requires users
to refer to the velvet manual. Also fixed a bug in line 233 and  line
208 and 209.

Also added a loop option to run velvetg for each kmer value computed
and to add their results to a single file for easier comparison and
selection for further tests.